+++ src/Entity/Metadata/Facet/ShopFacet.php
@@ -32,7 +32,7 @@ class ShopFacet extends EntityAbstract
     /**
      * @var bool
      */
-    protected $is_filter;
+    protected $is_filter = false;
 
     /**
      * @param int $id